Moved custom attributes to their own files and made them separate from the standard file.  Version 1 of this update.
