+++ holocron_api/api/models/LOB.py
@@ -1,0 +1,12 @@
+from django.db import models
+from standard import Placement
+
+class LOB(models.Model):
+    lob_key = models.SlugField(primary_key=True)
+    lob_description = models.CharField(max_length=140)
+    created_on = models.DateTimeField(auto_now_add=True)
+    updated = models.DateTimeField(auto_now=True)
+    placements = models.ManyToManyField(Placement)
+
+    def __str__(self):
+        return self.lob_key
